Refactor(#56): Repetiton 객체 분리
