+++ domain/jpa-calendar-domain/src/main/java/com/wypl/jpacalendardomain/schedule/data/ConvertibleSchedule.java
@@ -3,6 +3,8 @@ package com.wypl.jpacalendardomain.schedule.data;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
+import com.wypl.jpacalendardomain.schedule.domain.embedded.RepetitionCycle;
+
 public interface ConvertibleSchedule {
 
     String getTitle();
