--- domain/jpa-calendar-domain/src/main/java/com/wypl/jpacalendardomain/schedule/data/RepetitionCycle.java
@@ -1,8 +1,0 @@
-package com.wypl.jpacalendardomain.schedule.data;
-
-public enum RepetitionCycle {
-    DAY,
-    WEEK,
-    MONTH,
-    YEAR
-}
