+++ domain/jpa-calendar-domain/src/main/java/com/wypl/jpacalendardomain/schedule/domain/embedded/RepetitionCycle.java
@@ -1,0 +1,8 @@
+package com.wypl.jpacalendardomain.schedule.domain.embedded;
+
+public enum RepetitionCycle {
+	DAY,
+	WEEK,
+	MONTH,
+	YEAR
+}
